+++ CHANGELOG.md
@@ -57,6 +57,10 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
     This is supported for inputs and attributes.
 - Add `terramate.stack.parent.tags` metadata to stacks that are part of a parent-child hierarchy.
 
+### Changed
+
+- Local bundles and component definitions are no longer limited to the `/bundles` and `/components` root, but can be anywhere in the project.
+
 ## 0.16.0
 
 ### Merged from Terramate Catalyst